Axe `SearchableMenuItems.h`, moving things to `Menu.cpp` or `Menu.h` as necessary. Add `MenuTypes.h` for enums, structs, etc, for easy inclusion. Change `BenMenu` to `Menu`, and move to `Ship::` namespace rather than `BenGui::`.
