+++ mm/2s2h/BenGui/BenGui.hpp
@@ -8,7 +8,6 @@
 #include "DeveloperTools/CollisionViewer.h"
 #include "DeveloperTools/EventLog.h"
 #include "BenInputEditorWindow.h"
-#include "Menu.h"
 
 namespace BenGui {
     void SetupHooks();
